Refactoring SaslServerCallbackHandler and SaslClientCallbackHandler
